Handle short reads in the local sockets example
Protection against short reads was already half implemented,
blockSize was being sent by the server but never used by the client.

Also, blockSize was bumped to quint32: If you're in a position where
short reads can happen then quint16 is probably not enough to hold the
size of your data. On Linux I could only reproduce short reads for
messages > 500K.

+++ examples/corelib/ipc/localfortuneserver/server.cpp
@@ -96,10 +96,10 @@ void Server::sendFortune()
     QByteArray block;
     QDataStream out(&block, QIODevice::WriteOnly);
     out.setVersion(QDataStream::Qt_4_0);
-    out << (quint16)0;
+    out << (quint32)0;
     out << fortunes.at(qrand() % fortunes.size());
     out.device()->seek(0);
-    out << (quint16)(block.size() - sizeof(quint16));
+    out << (quint32)(block.size() - sizeof(quint32));
 
     QLocalSocket *clientConnection = server->nextPendingConnection();
     connect(clientConnection, SIGNAL(disconnected()),
